Added FactorioProcess class
