+++ src/Factoriod.Daemon/FactorioHostedService.cs
@@ -1,5 +1,3 @@
-using System.Diagnostics;
-using System.Text.RegularExpressions;
 using Factoriod.Fetcher;
 using Factoriod.Models;
 using Microsoft.Extensions.Hosting;
@@ -15,15 +13,17 @@ namespace Factoriod.Daemon
         private readonly IHostApplicationLifetime lifetime;
         private readonly VersionFetcher versionFetcher;
         private readonly ReleaseFetcher releaseFetcher;
+        private readonly FactorioProcess factorioProcess;
 
         public FactorioHostedService(ILogger<FactorioHostedService> logger, IOptions<Options.Factorio> options,
-            IHostApplicationLifetime lifetime, VersionFetcher versionFetcher, ReleaseFetcher releaseFetcher)
+            IHostApplicationLifetime lifetime, VersionFetcher versionFetcher, ReleaseFetcher releaseFetcher, FactorioProcess factorioProcess)
         {
             this.logger = logger;
             this.options = options.Value;
             this.lifetime = lifetime;
             this.versionFetcher = versionFetcher;
             this.releaseFetcher = releaseFetcher;
+            this.factorioProcess = factorioProcess;
         }
 
         protected override async Task ExecuteAsync(CancellationToken cancellationToken)
@@ -45,7 +45,7 @@ namespace Factoriod.Daemon
                 return;
             }
 
-            var exitCode = await StartServerAsync(factorioDirectory, cancellationToken);
+            var exitCode = await this.factorioProcess.StartServerAsync(factorioDirectory, cancellationToken);
             if (!cancellationToken.IsCancellationRequested && exitCode != 0)
             {
                 this.logger.LogError("Factorio process exited early with code {exitCode}, shutting down", exitCode);
@@ -116,180 +116,6 @@ namespace Factoriod.Daemon
 
             this.logger.LogDebug("Downloaded Factorio version {version} to {path}", requestedVersion, downloadedDirectory);
             return downloadedDirectory;
-        }
-
-        private async Task<int> StartServerAsync(DirectoryInfo factorioDirectory, CancellationToken cancellationToken = default)
-        {
-            this.logger.LogInformation("Starting factorio process");
-
-            var arguments = new List<string>();
-            var addedStartServer = AddArgumentIfFileExists(arguments, "--start-server", this.options.Saves.GetSavePath());
-            if (!addedStartServer)
-            {
-                this.logger.LogInformation("No save file found, creating one");
-                await CreateSaveIfNotExists(factorioDirectory, cancellationToken);
-
-                // try again
-                addedStartServer = AddArgumentIfFileExists(arguments, "--start-server", this.options.Saves.GetSavePath());
-                if (!addedStartServer)
-                {
-                    this.logger.LogError("Unable to find save file {path}", this.options.Saves.GetSavePath());
-                    return 1;
-                }
-            }
-
-            AddArgumentIfFileExists(arguments, "--server-settings", this.options.Configuration.GetServerSettingsPath());
-            var addedWhitelist = AddArgumentIfFileExists(arguments, "--server-whitelist", this.options.Configuration.GetServerWhitelistPath());
-            if (addedWhitelist)
-            {
-                arguments.Add("--use-server-whitelist");
-            }
-
-            AddArgumentIfFileExists(arguments, "--server-banlist", this.options.Configuration.GetServerBanlistPath());
-            AddArgumentIfFileExists(arguments, "--server-adminlist", this.options.Configuration.GetServerAdminlistPath());
-            AddArgumentIfDirectoryExists(arguments, "--mod-directory", this.options.GetModsRootDirectory());
-
-            using var factorioProcess = new Process()
-            {
-                StartInfo = new ProcessStartInfo
-                {
-                    FileName = Path.Combine(factorioDirectory.FullName, this.options.Executable.ExecutableDirectory, this.options.Executable.ExecutableName),
-                    Arguments = string.Join(" ", arguments),
-                    WorkingDirectory = factorioDirectory.FullName,
-                    UseShellExecute = false,
-                    RedirectStandardOutput = true,
-                    RedirectStandardError = true,
-                    RedirectStandardInput = true,
-                    CreateNoWindow = true,
-                },
-            };
-
-            await StartProcessWithOutputHandlersAndWaitForExitAsync(factorioProcess, cancellationToken);
-            return factorioProcess.ExitCode;
-        }
-
-        private async Task CreateSaveIfNotExists(DirectoryInfo factorioDirectory, CancellationToken cancellationToken = default)
-        {
-            var savePath = this.options.Saves.GetSavePath();
-            if (savePath.Exists)
-            {
-                return;
-            }
-
-            if (cancellationToken.IsCancellationRequested)
-            {
-                return;
-            }
-
-            this.logger.LogInformation("Creating save file {path}", savePath);
-
-            var arguments = new List<string>()
-            {
-                "--create",
-                savePath.FullName,
-            };
-
-            AddArgumentIfFileExists(arguments, "--map-gen-settings", this.options.MapGeneration.GetMapGenSettingsPath());
-            AddArgumentIfFileExists(arguments, "--map-settings", this.options.MapGeneration.GetMapSettingsPath());
-            if (this.options.MapGeneration.MapGenSeed.HasValue)
-            {
-                arguments.Add("--map-gen-seed");
-                arguments.Add(this.options.MapGeneration.MapGenSeed.Value.ToString());
-            }
-            
-            using var createSaveProcess = new Process
-            {
-                StartInfo = new ProcessStartInfo
-                {
-                    FileName = this.options.Executable.GetExecutablePath(factorioDirectory).FullName,
-                    Arguments = string.Join(" ", arguments),
-                    WorkingDirectory = factorioDirectory.FullName,
-                    UseShellExecute = false,
-                    RedirectStandardOutput = true,
-                    RedirectStandardError = true,
-                    RedirectStandardInput = true,
-                    CreateNoWindow = true,
-                },
-            };
-
-            await StartProcessWithOutputHandlersAndWaitForExitAsync(createSaveProcess, cancellationToken);
-        }
-
-        private static bool AddArgumentIfFileExists(List<string> arguments, string option, FileInfo path)
-        {
-            if (path.Exists)
-            {
-                arguments.Add(option);
-                arguments.Add(path.FullName);
-                return true;
-            }
-
-            return false;
-        }
-
-        private static bool AddArgumentIfDirectoryExists(List<string> arguments, string option, DirectoryInfo path)
-        {
-            if (path.Exists)
-            {
-                arguments.Add(option);
-                arguments.Add(path.FullName);
-                return true;
-            }
-
-            return false;
-        }
-
-        private void OnFactorioProcessOutputDataReceived(object sender, DataReceivedEventArgs e)
-        {
-            if (e.Data is null)
-            {
-                return;
-            }
-
-            this.logger.LogTrace("Factorio process output: {output}", e.Data);
-            var badVersionMatch = Regex.Match(e.Data, @"Map version (?<new_version>\d+\.\d+\.\d+)-0 cannot be loaded because it is higher than the game version \((?<old_version>\d+\.\d+\.\d+)-0\)");
-            if (badVersionMatch.Success)
-            {
-                // TODO(#24): Handle newer maps being loaded by older server versions
-                this.logger.LogWarning("Factorio map version {new_version} cannot be loaded because it is higher than the game version {old_version}", badVersionMatch.Groups["new_version"].Value, badVersionMatch.Groups["old_version"].Value);
-            }
-
-            if (e.Data.Contains("changing state from(CreatingGame) to(InGame)"))
-            {
-                this.logger.LogInformation("Factorio process started");
-            }
-        }
-
-        private void OnFactorioProcessErrorDataReceived(object sender, DataReceivedEventArgs e)
-        {
-            if (e.Data is null)
-            {
-                return;
-            }
-
-            this.logger.LogWarning("Factorio process error: {error}", e.Data);
-        }
-
-        private async Task StartProcessWithOutputHandlersAndWaitForExitAsync(Process process, CancellationToken cancellationToken = default)
-        {
-            if (cancellationToken.IsCancellationRequested)
-            {
-                return;
-            }
-            
-            process.OutputDataReceived += this.OnFactorioProcessOutputDataReceived;
-            process.ErrorDataReceived += this.OnFactorioProcessErrorDataReceived;
-
-            process.Start();
-
-            process.BeginOutputReadLine();
-            process.BeginErrorReadLine();
-
-            // TODO(#25): something is causing the last of the output to be lost and not sent to OnFactorioProcessOutputDataReceived
-            await process.WaitForExitAsync(cancellationToken);
-
-            process.CancelOutputRead();
-            process.CancelErrorRead();
         }
     }
 }
